requirements.txt
arrow==0.8.0
coreapi==2.0.8
Django==1.10.3
django-cors-middleware==1.3.1
django-filter==0.15.3
django-redis==4.5.0
django-rest-swagger==2.0.7
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
itypes==1.1.0
Markdown==2.6.7
MySQL-python==1.2.5
openapi-codec==1.1.5
Pillow==3.4.2
PyJWT==1.4.2
python-dateutil==2.5.3
python-memcached==1.58
redis==2.10.5
requests==2.11.1
simplejson==3.8.2
six==1.10.0
typing==3.5.2.2
uritemplate==3.0.0
bleach==1.5.0
